+++ lua/plugins/config/syntax_tree_surfer.lua
@@ -5,20 +5,20 @@ return function(use)
 		config = function()
 			local wk = require('which-key')
 
-			wk.register({
-				['<c-n>'] = {
-					[[<cmd>lua require('syntax-tree-surfer').select_current_node()<cr>]],
-					'Select Current Node',
-				},
-				['<c-m>'] = { [[<cmd>lua require('syntax-tree-surfer').select()<cr>]], 'Select Preview' },
-			})
+			-- wk.register({
+			-- 	['<c-n>'] = {
+			-- 		[[<cmd>lua require('syntax-tree-surfer').select_current_node()<cr>]],
+			-- 		'Select Current Node',
+			-- 	},
+			-- 	['<c-m>'] = { [[<cmd>lua require('syntax-tree-surfer').select()<cr>]], 'Select Preview' },
+			-- })
 
 			wk.register({
 				-- Navigation
 				['.'] = { [[<cmd>lua require('syntax-tree-surfer').surf("prev", "visual")<cr>]], 'Surf Previous' },
 				[','] = { [[<cmd>lua require('syntax-tree-surfer').surf("next", "visual")<cr>]], 'Surf Next' },
-				['<c-m>'] = { [[<cmd>lua require('syntax-tree-surfer').surf("parent", "visual")<cr>]], 'Surf Parent' },
-				['<c-n>'] = { [[<cmd>lua require('syntax-tree-surfer').surf("child", "visual")<cr>]], 'Surf Child' },
+				-- ['<c-m>'] = { [[<cmd>lua require('syntax-tree-surfer').surf("parent", "visual")<cr>]], 'Surf Parent' },
+				-- ['<c-n>'] = { [[<cmd>lua require('syntax-tree-surfer').surf("child", "visual")<cr>]], 'Surf Child' },
 
 				-- Swapping
 				['<c-h>'] = {