+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 > Features
 - **go:** changed from ray-x-go to personal custom one
+- **surfer:** more sane keybinding
 
 
 <a name="v4.0.0"></a>