+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 > Features
 - **go:** changed from ray-x-go to personal custom one
+- **iswap:** added iswap
 - **surfer:** more sane keybinding
 
 > Various Actions