+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - better grammar desc on TextYankPost
 
 > Features
+- **cmp:** added source on completion
 - **go:** changed from ray-x-go to personal custom one
 - **iswap:** added iswap
 - **nvim-cmp:** add limit on many source hits like buffer, rg, spell