+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - **debug-neovim:** move command to lua
 - **go:** changed from ray-x-go to personal custom one
 - **iswap:** added iswap
+- **neovide:** changed font to comic code liagures
 - **nvim-cmp:** add limit on many source hits like buffer, rg, spell
 - **settings:** moved to lua
 - **sumneko:** removed lua-dev in favor of cmp-nvim-lua and removed buggy ui-select from telescope