+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - better grammar desc on TextYankPost
 
 > Features
+- **catppuccin:** added more integration
 - **cmp:** added source on completion
 - **debug-neovim:** move command to lua
 - **go:** changed from ray-x-go to personal custom one