+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 > Bug Fixes
 - **apm-span-snippet:** fix treesitter query
 - **treesitter:** finish mapping ts_utils.get_node_text to vim.treesitter.query.get_node_text
+- **which-key:** fix wtf mapping for buffer
 
 > Code Refactoring
 - checktime now is silent!